Add files via upload
New Features:
Both Doctors and Secretaries are now able to create appointments for patients. Both are not able to create duplicate appointments for the same date/time. 
A Patient is now able to register for a new account.
An admin is able to create a new account, and is able to create new doctors and secretaries.
A Doctor is able to see their upcoming appointments. When an appointment is clicked on, the Doctor is able to make notes regarding the appointment, and is able to prescribe new medicines. If the medicine they wish to prescribe does not exist, the Doctor is able to create said medicine, and request it to be ordered. 
A Secretary is able to see all medicines and patients, and search both by name.
The Secretary is able to dispense prescriptions (if medicine stock is available)
The Secretary is able to modify a Medicine.
A Secretary is able to approve Patient accounts, and change the status of an account, on request.