Add files via upload
Fixed error in previous version, which combined the UID and Name into a string which was then used for filtering different users by type. This resulted in all users with a "P" in their name being included in a list of patients, or patients whose name included the letter "D" being included in the list of doctors. 
To fix this, I created a new IFilterByStringBehaviour concrete class FilterByUID, which specifically filters by UID, and explicitly used this behaviour when filtering by UID is required.
In implementations of the ListableByString function getSearchableString(), i also removed the +UID from the return string. Although combining the UID and name might be useful for allowing the use of the same behaviour to search for UID's or a name, it might cause unexpected or misleading results.