Add files via upload
Fixed NullPointerException that was thrown in User.getUID(). file.ListOfFiles() would return null if the directory did not exist. To solve this, i added code in the Login.main to create the missing directories /users and /medicines on startup. Using .mkdirs() will do nothing if the folders already exist. I also changed the conditional in User.getUID() to break the loop if the listOfFiles is null. 
Additionally, I added a function to the Appointments class to get the Appointments' notes, allowing the AppointmentWindow to display the notes. 
Finally, i added labels to show the newly created UID's.